requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
executing==2.2.1
fonttools==4.61.1
fpdf2==2.8.5
gensim==4.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jieba==0.42.1
jiter==0.12.0
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
openai==2.16.0
opencv-python==4.13.0.90
packaging==26.0
pandas==3.0.0
parso==0.8.5
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
PyGetWindow==0.0.9
Pygments==2.19.2
pynput==1.8.1
pyparsing==3.3.2
PyRect==0.2.0
python-dateutil==2.9.0.post0
pyzmq==27.1.0
requests==2.32.5
scikit-image==0.26.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
stack-data==0.6.3
tifffile==2026.1.28
tornado==6.5.4
tqdm==4.67.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.5.3
wordcloud==1.9.6
wrapt==2.1.1
